NIAD-509 MockSpine mode added to SCR app
