+++ docker/service/src/main/java/uk/nhs/adaptors/scr/config/ScrConfiguration.java
@@ -13,4 +13,5 @@ public class ScrConfiguration {
     private String partyIdTo;
     private String partyIdFrom;
     private String nhsdAsidTo;
+    private ScrSpineMode spineMode;
 }
