+++ docker/service/src/main/java/uk/nhs/adaptors/scr/config/ScrSpineMode.java
@@ -1,0 +1,6 @@
+package uk.nhs.adaptors.scr.config;
+
+public enum ScrSpineMode {
+    REAL,
+    MOCK;
+}
