+++ docker/service/src/main/resources/application.yml
@@ -21,6 +21,7 @@ management:
         include: info, health, metrics, mappings
 
 scr:
+  spineMode: ${SCR_SPINE_MODE:REAL}
   partyIdFrom: ${SCR_PARTY_ID_FROM:TESTGEN-201324}
   partyIdTo: ${SCR_PARTY_ID_TO}
   nhsdAsidTo: ${SCR_NHSD_ASID_TO}
