+++ ecs-proxies-containers.yml
@@ -3,5 +3,5 @@ docker_containers:
     dockerfile: "docker/service/Dockerfile"
     path: "docker/service"
   - name: scr-sandbox
-    dockerfile: "docker/sandbox/Dockerfile"
-    path: "docker/sandbox"
+    dockerfile: "docker/service/Dockerfile"
+    path: "docker/service"
