+++ ecs-proxies-deploy.yml
@@ -10,6 +10,8 @@ docker_service:
           value: "{{ 'DEBUG' if APIGEE_ENVIRONMENT == 'internal-dev' else 'INFO' }}"
         - name: SCR_SPINE_URL
           value: "https://msg.intspineservices.nhs.uk"
+        - name: SCR_SPINE_MODE
+          value: "REAL"
       secrets:
         - name: SCR_SPINE_ENDPOINT_KEY
           valueFrom: "/ptl/api-deployment/scr/ldap/private-key"
