+++ proxies/sandbox/apiproxy/policies/AssignMessage.AddBaseUrl.xml
@@ -1,0 +1,11 @@
+<AssignMessage async="false" continueOnError="false" enabled="true" name="AssignMessage.AddBaseUrl">
+    <FaultRules/>
+    <Properties/>
+    <Set>
+        <Headers>
+          <Header name="client-request-url">{proxy.url}</Header>
+        </Headers>
+    </Set>
+    <IgnoreUnresolvedVariables>true</IgnoreUnresolvedVariables>
+    <AssignTo createNew="false" transport="http" type="response"/>
+</AssignMessage>
